Trait mincodec::Deserialize

pub trait Deserialize: Sized {
    type Target;
    type Error;
    fn poll_deserialize<B: BitBuf>(
        self: Pin<&mut Self>, 
        ctx: &mut Context, 
        buf: B
    ) -> BufPoll<Result<Self::Target, Self::Error>>;
}

Describes the manner in which a type is deserialized from a bit-level buffer

Associated Types

type Target

The type this deserializer produces

type Error

The error type produced on failure of the deserialization process

Loading content...
Required methods

fn poll_deserialize<B: BitBuf>(
self: Pin<&mut Self>,
ctx: &mut Context,
buf: B
) -> BufPoll<Result<Self::Target, Self::Error>>

Advances the deserialization process. This method should write to the buffer using partial methods or Drain/CappedDrain (i.e. not write_all) and use the provided helper macros to return BufPoll::Insufficient when the buffer is full or Poll::Pending when a future is pending and will wakeup using ctx.

Loading content...