Trait mincodec::Serialize

pub trait Serialize {
    type Error;
    fn poll_serialize<B: BitBufMut>(
        self: Pin<&mut Self>, 
        ctx: &mut Context, 
        buf: B
    ) -> BufPoll<Result<(), Self::Error>>;
}

Describes the manner in which a type is serialized into a bit-level buffer

Associated Types

type Error

The error type produced on failure of the serialization process

Loading content...
Required methods

fn poll_serialize<B: BitBufMut>(
self: Pin<&mut Self>,
ctx: &mut Context,
buf: B
) -> BufPoll<Result<(), Self::Error>>

Advances the serialization process. This method should read from the buffer using partial methods or Fill/CappedFill (i.e. not read_all) and use the provided helper macros to return BufPoll::Insufficient when more data is needed or Poll::Pending when a future is pending and will wakeup using ctx.

Loading content...